Removed singleton pattern

+++ 3-semester/Programming/Week 16/SimpleRestExercise/SimpleRestExercise/Utilities/FileReader.cs
@@ -4,26 +4,9 @@ namespace SimpleRestExercise.Utilities;
 
 public class FileReader<T>
 {
-    private static FileReader<T>? instance = null;
-    private static readonly object padlock = new object();
     private static readonly string baseDirectory = AppDomain.CurrentDomain.BaseDirectory;
     private static readonly string relativeFolderPath = "Data";
 
-    public static FileReader<T>? Instance
-    {
-        get
-        {
-            lock (padlock)
-            {
-                if (instance == null)
-                {
-                    instance = new FileReader<T>();
-                }
-
-                return instance;
-            }
-        }
-    }
 
     public async Task Save(string fileName, T data)
     {
@@ -34,7 +17,10 @@ public class FileReader<T>
 
         string filePath = Path.Combine(baseDirectory, relativeFolderPath, fileName);
 
-        Directory.CreateDirectory(Path.GetDirectoryName(filePath));
+        if (!File.Exists(filePath))
+        {
+            throw new FileNotFoundException($"{filePath} does not exist");
+        }
 
         await using FileStream createStream = File.Create(filePath);
         await JsonSerializer.SerializeAsync(createStream, data);
